requirements.txt
# Existing requirements
boto3>=1.20.0
sagemaker>=2.80.0
pillow>=8.4.0
tensorboard>=2.8.0
matplotlib>=3.5.0
tqdm>=4.62.0
pytorch-fid>=0.2.0

# SageMaker specific
sagemaker-inference>=1.5.0
sagemaker-training>=3.6.0